% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure.R
\name{create_structure}
\alias{create_structure}
\title{Create research project folder structure}
\usage{
create_structure(pj_structure = pj_structure, root = NULL)
}
\arguments{
\item{pj_structure}{a alist of folder structure, default = pj_structure in package data}

\item{root}{string of root absolute path, default = NULL meaning project root}
}
\value{
creating folder structure
}
\description{
Create research project folder structure
}
