% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_questions.R
\name{format_questions}
\alias{format_questions}
\title{Format questions to prepare for export}
\usage{
format_questions(.data, language, raw = FALSE)
}
\arguments{
\item{.data}{A qsf dataframe}

\item{language}{language of survey}

\item{raw}{raw includes Answers Class and no spaces; default set to FALSE}
}
\value{
formatted dataframe ready for export
}
\description{
Format questions to prepare for export
}
