% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fre.R
\name{fre}
\alias{fre}
\title{Get labels from a dataset}
\usage{
fre(.data, var)
}
\arguments{
\item{.data}{A dataset.}

\item{var}{A variable.}
}
\value{
The frequency of a variable.
}
\description{
Get labels from a dataset
}
