% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLFM.R
\name{glfm}
\alias{glfm}
\title{Generalized linear function regression model}
\usage{
glfm(
  Y,
  X,
  Z = NULL,
  q = 4,
  mxex.smooth = FALSE,
  sort.mx = "CR-direct",
  n.order.beta = 4,
  n.basis.beta = 7,
  base.system.beta = "B-spline",
  n.order.mxex = 4,
  n.basis.mxex = 7,
  base.system.mxex = "B-spline",
  family = "gaussian",
  screening = TRUE
)
}
\arguments{
\item{Y}{a numeric vector of response, which can be continuous or binary}

\item{X}{a numeric matrix of mixture exposure}

\item{Z}{a data frame of covariates, default value is NULL for no covariates model}

\item{q}{a number to decide quantile setting}

\item{sort.mx}{logical value to decide if smoothing function for exposure is needed}

\item{n.order.beta}{order of expansion model for mixture effect function}

\item{n.basis.beta}{number of basis functions for mixture effect function}

\item{base.system.beta}{basis function system for mixture effect function}

\item{n.order.mxex}{order of expansion model for mixture exposure function}

\item{n.basis.mxex}{number of basis functions for mixture exposure function}

\item{base.system.mxex}{basis function system for mixture exposure function}

\item{family}{specify link function, options are "gaussian" and "binomial"}

\item{screening}{logical value to specify outputs, if TRUE, mixture effect function is estimated}

\item{clustering}{logical value to decide if hierarchical clustering is needed for exposure matrix to sort variables in the way that correlated ones are closer}
}
\value{
Not Applicable
}
\description{
To fit generalized linear function regression model
}
\examples{
Y=y;X=x;Z=z;q=4;sort.mx="ASC";mxex.smooth = FALSE;n.order.beta=4;n.basis.beta=7;base.system.beta="B-spline";n.order.mxex=4;n.basis.mxex=7;base.system.mxex="B-spline";family="binomial"
}
