% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{labels}
\alias{labels}
\title{Get labels from a dataset}
\usage{
labels(.data)
}
\arguments{
\item{.data}{A dataset.}
}
\value{
The labels of a dataset.
}
\description{
Get labels from a dataset
}
\examples{
labels(mtcars)
}
