% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotsums.R
\name{plotsums}
\alias{plotsums}
\title{test}
\usage{
plotsums(a = 2, b = 3)
}
\arguments{
\item{test}{test}
}
\value{
test
}
\description{
test
}
\details{
missing
}
\examples{
missing
}
\references{
Marvin Reich (2017), mreich@posteo.de
}
