% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sums.R
\name{sums}
\alias{sums}
\title{test}
\usage{
sums(a = 1, b = 2)
}
\arguments{
\item{test}{test}
}
\value{
test
}
\description{
test
}
\details{
missing
}
\examples{
missing
}
\references{
Marvin Reich (2017), mreich@posteo.de
}
